+++ src/main/java/com/microsoft/spring/data/gremlin/conversion/result/GremlinResultReader.java
@@ -3,8 +3,9 @@
  * Licensed under the MIT License. See LICENSE in the project root for
  * license information.
  */
-package com.microsoft.spring.data.gremlin.conversion;
+package com.microsoft.spring.data.gremlin.conversion.result;
 
+import com.microsoft.spring.data.gremlin.conversion.source.GremlinSource;
 import org.apache.tinkerpop.gremlin.driver.Result;
 
 public interface GremlinResultReader {
