+++ src/main/java/com/microsoft/spring/data/gremlin/conversion/script/GremlinScriptPropertiesLiteral.java
@@ -3,7 +3,7 @@
  * Licensed under the MIT License. See LICENSE in the project root for
  * license information.
  */
-package com.microsoft.spring.data.gremlin.conversion;
+package com.microsoft.spring.data.gremlin.conversion.script;
 
 import com.microsoft.spring.data.gremlin.common.Constants;
 import org.apache.commons.lang3.NotImplementedException;
