+++ src/main/java/com/microsoft/spring/data/gremlin/conversion/script/GremlinScriptVertexFindByIdLiteral.java
@@ -3,9 +3,11 @@
  * Licensed under the MIT License. See LICENSE in the project root for
  * license information.
  */
-package com.microsoft.spring.data.gremlin.conversion;
+package com.microsoft.spring.data.gremlin.conversion.script;
 
 import com.microsoft.spring.data.gremlin.common.Constants;
+import com.microsoft.spring.data.gremlin.conversion.source.GremlinSource;
+import com.microsoft.spring.data.gremlin.conversion.source.GremlinSourceVertex;
 import com.microsoft.spring.data.gremlin.exception.UnexpectedGremlinSourceTypeException;
 import lombok.NoArgsConstructor;
 import org.springframework.lang.NonNull;
@@ -13,10 +15,12 @@ import org.springframework.util.Assert;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 
+/**
+ * Provide generateScript findById from GremlinSourceVertex.
+ */
 @NoArgsConstructor
-public class GremlinScriptVertexLiteral extends GremlinScriptPropertiesLiteral implements GremlinScript<String> {
+public class GremlinScriptVertexFindByIdLiteral implements GremlinScript<String> {
 
     @Override
     public String generateScript(@NonNull GremlinSource source) {
@@ -25,20 +29,12 @@ public class GremlinScriptVertexLiteral extends GremlinScriptPropertiesLiteral i
         }
 
         final List<String> scriptList = new ArrayList<>();
-        final String label = source.getLabel();
         final String id = source.getId();
-        final Map<String, Object> properties = source.getProperties();
 
-        Assert.notNull(label, "label should not be null");
         Assert.notNull(id, "id should not be null");
-        Assert.notNull(properties, "properties should not be null");
 
         scriptList.add(Constants.GREMLIN_PRIMITIVE_GRAPH);
-
-        scriptList.add(String.format(Constants.GREMLIN_PRIMITIVE_ADD_VERTEX, label));
-        scriptList.add(String.format(Constants.GREMLIN_PRIMITIVE_PROPERTY_STRING, Constants.PROPERTY_ID, id));
-
-        super.generateGremlinScriptProperties(scriptList, properties);
+        scriptList.add(String.format(Constants.GREMLIN_PRIMITIVE_VERTEX, id));
 
         return String.join(Constants.GREMLIN_PRIMITIVE_INVOKE, scriptList);
     }
