+++ src/main/java/com/microsoft/spring/data/gremlin/conversion/source/BasicGremlinSourceWriter.java
@@ -3,8 +3,9 @@
  * Licensed under the MIT License. See LICENSE in the project root for
  * license information.
  */
-package com.microsoft.spring.data.gremlin.conversion;
+package com.microsoft.spring.data.gremlin.conversion.source;
 
+import com.microsoft.spring.data.gremlin.conversion.MappingGremlinConverter;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.lang.NonNull;
