+++ src/main/java/com/microsoft/spring/data/gremlin/conversion/source/GremlinSourceReader.java
@@ -3,7 +3,9 @@
  * Licensed under the MIT License. See LICENSE in the project root for
  * license information.
  */
-package com.microsoft.spring.data.gremlin.conversion;
+package com.microsoft.spring.data.gremlin.conversion.source;
+
+import com.microsoft.spring.data.gremlin.conversion.MappingGremlinConverter;
 
 /**
  * Provider Entity type dependent read method.
