+++ src/main/java/com/microsoft/spring/data/gremlin/conversion/source/GremlinSourceVertex.java
@@ -3,7 +3,7 @@
  * Licensed under the MIT License. See LICENSE in the project root for
  * license information.
  */
-package com.microsoft.spring.data.gremlin.conversion;
+package com.microsoft.spring.data.gremlin.conversion.source;
 
 import org.springframework.lang.NonNull;
