+++ src/main/java/com/microsoft/spring/data/gremlin/repository/support/GremlinEntityInformation.java
@@ -10,9 +10,13 @@ import com.microsoft.spring.data.gremlin.annotation.Graph;
 import com.microsoft.spring.data.gremlin.annotation.Vertex;
 import com.microsoft.spring.data.gremlin.common.Constants;
 import com.microsoft.spring.data.gremlin.common.GremlinEntityType;
+import com.microsoft.spring.data.gremlin.conversion.script.GremlinScript;
+import com.microsoft.spring.data.gremlin.conversion.script.GremlinScriptEdgeLiteral;
+import com.microsoft.spring.data.gremlin.conversion.script.GremlinScriptGraphLiteral;
+import com.microsoft.spring.data.gremlin.conversion.script.GremlinScriptVertexLiteral;
+import com.microsoft.spring.data.gremlin.conversion.source.*;
 import com.microsoft.spring.data.gremlin.exception.InvalidGremlinEntityIdFieldException;
 import com.microsoft.spring.data.gremlin.exception.UnexpectedGremlinEntityTypeException;
-import com.microsoft.spring.data.gremlin.conversion.*;
 import org.apache.commons.lang3.reflect.FieldUtils;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.repository.core.support.AbstractEntityInformation;
